+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 - It possible to easily specify expecting that it equals to specified characters.
   - Please refer to a class that name is `MAERawFragment`.
-- Add `MAEArrayAdapter # valueByFragmentWithFormat:separatedStrings:error:`
-  - Implement of `classForParsingArray` will easier.
+- Implement of `classForParsingArray` will easier.
+  - `MAEArrayAdapter # valueByFragmentWithFormat:separatedStrings:error:`
+  - `MAEArrayAdapter # defaultClassForParsingArray:withCandidateClasses:`
 
 ### Major changes